Added JUnitTest for RemoveCommand and DeleteCommand
